Added findPtIndexes and applyForce utility functions
Added function findPtIndexes which takes the point ID values from a line and uses the find function to identify the corresponding index in the points matrix.
Added the applyForce function which adds the sum of a force and a unit vector to the x and y components of a point's force value.